Added Model + Refactor
