+++ app/src/main/java/com/example/gocoronago/tipspage/StayHomeDataFragment.kt
@@ -1,4 +1,4 @@
-package com.example.gocoronago.stayHome
+package com.example.gocoronago.tipspage
 
 import android.os.Bundle
 import android.view.LayoutInflater
