+++ app/src/main/java/com/example/gocoronago/tipspage/viewmodel/StayHomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.gocoronago.tipspage.viewmodel
+
+import dagger.hilt.android.lifecycle.HiltViewModel
+
+@HiltViewModel
+class StayHomeViewModel {
+}
