+++ app/src/main/java/com/example/gocoronago/ui/main/MainFragment.kt
@@ -16,7 +16,7 @@ import com.example.gocoronago.R
 import com.example.gocoronago.base.BaseFragment
 import com.example.gocoronago.base.RequestResult
 import com.example.gocoronago.databinding.MainFragmentBinding
-import com.example.gocoronago.stayHome.StayHomeActivity
+import com.example.gocoronago.tipspage.StayHomeActivity
 import kotlinx.android.synthetic.main.main_fragment.*
 import kotlinx.android.synthetic.main.stay_home.*
 import kotlinx.android.synthetic.main.total_cases.*
